ai/requirements.txt
fastapi
uvicorn
pydantic
typing-extensions
google-generativeai
python-dotenv
